+++ moses.lua
@@ -384,15 +384,33 @@ function M.cycle(t, n)
   end
 end
 
---- Maps `f (v, k)` on value-key pairs, collects and returns the results.
+--- Maps `f (v, k)` on value-key pairs, collects and returns the results. 
+-- Uses `pairs` to iterate over elements in `t`.
 -- <br/><em>Aliased as `collect`</em>.
 -- @name map
 -- @param t a table
 -- @param f  an iterator function, prototyped as `f (v, k)`
 -- @return a table of results
+-- @see mapi
 function M.map(t, f)
   local _t = {}
   for index,value in pairs(t) do
+    local k, kv, v = index, f(value, index)
+    _t[v and kv or k] = v or kv
+  end
+  return _t
+end
+
+--- Maps `f (v, k)` on value-key pairs, collects and returns the results. 
+-- Uses `ipairs` to iterate over elements in `t`.
+-- @name mapi
+-- @param t a table
+-- @param f  an iterator function, prototyped as `f (v, k)`
+-- @return a table of results
+-- @see map
+function M.mapi(t, f)
+  local _t = {}
+  for index,value in ipairs(t) do
     local k, kv, v = index, f(value, index)
     _t[v and kv or k] = v or kv
   end